SAA: Fix permission display with more than one embedded permission
The Android Preference library doesn't like it when the same key is
assigned to multiple preferences. It should be unique. This is
causing issues with embedded permissions and potentially also chooser
permissions.
Fix the issue by keeping the preferences in a list instead of
assigning keys to them.
Also ensure that the icon of embedded permissions is updated
when a permission is blocked.

Low-Coverage-Reason: TESTS_IN_SEPARATE_CL Chooser permissions are not covered by tests. I'll add tests separately since this is a little complicated from Java.
Bug: 1497948
Change-Id: I10ed3436454d82612c42872bdad3abb5a970b915
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4993549
Commit-Queue: Christian Dullweber <dullweber@chromium.org>
Auto-Submit: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Finnur Thorarinsson <finnur@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218243}