ash: Fix LayerAnimationObserver crash during unlock
SessionStateAnimatorImpl::AnimationSequence could be destructed
prematurely when LayerCopyAnimator aborts animations due to container
bounds change during animation setup. On an edge case with multiple
displays, this could happen to containers in one display but not
the other. As a result, the animations that are not aborted would
hit use-after-free. This CL defers the animation completed signal
until EndSequence is called so that the animation sequence would be
kept around when there are pending animations.

Bug: b:252559173
Change-Id: I8612918d51bec4dbb5a51f8d1292fcb88df77035
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4995768
Reviewed-by: Sammie Quon <sammiequon@chromium.org>
Commit-Queue: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218336}