Fix erroneous excessive-trigger-data check in TriggerSpecs::Parse
Prior to this change, the number of *specs* seen so far was erroneously
used as the number of trigger data values seen so far. The correct count
is determined by the size of the `trigger_data_indices` map.

Note that TriggerSpecs::Parse is currently uncalled outside tests, so
the erroneous behavior was never web-visible.

Bug: 1495162
Change-Id: Ie9452cba7976ff6c40d09e0ce5d25c4b21a3606c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994735
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: Andrew Paseltiner <apaseltiner@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218342}