webdialogs: remove OnDialogClosingFromKeyEvent()
This method was only overridden in one place, where it was used to call
OnDialogClosed() unnecessarily - in that situation, the dialog is just
about to close anyway, so OnDialogClosed() is about to be run by
WebDialogDelegate itself. This change deletes the method, its only
override, and its only call site. Simplification of
PrintPreviewDialogController as a result might be possible, since it no
longer needs to tolerate multiple calls to OnDialogClosed().

This change does not add an automated test. I tried to do so, but:
* A unit test that doesn't depend on Views didn't work: there's no way
  to send key events to the dialog directly without Views APIs;
* A unit test that does depend on Views required putting a single unit
  test for this code, which otherwise doesn't use Views, in a different
  part of the source tree, which seems very ugly;
* A non-unit test that doesn't depend on Views would need to be an
  interactive ui test so that it could use ui_controls::SendKeyPress().
  Given the cost of interactive ui tests, how flaky they generally are,
  and the low risk of this code regressing in the future, this seems not
  worth writing.

Instead, I manually tested that I did not regress the behavior.

Bug: 1485127
Change-Id: I804d56adff25739618646f8c6ee926a4c371342a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994890
Auto-Submit: Elly FJ <ellyjones@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Scott Violet <sky@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218353}