Make Lens + companion one side panel entry when Lens is contextual.
When companion image search is disabled and Lens is a contextual panel, we want to use the "SearchCompanion" entry ID for both of the views. This allows for the pinned button to be highlighted when either of these views is showing and for the combobox deletion hack added in a previous CL to be removed. This also allows us to delete all the code in the lens coordinator since it no longer creates the view for the contextual Lens panel.

Bug: b:301638334, b:303842316
Change-Id: I5d5d67555cd008e9f032b8838f8f0de3de66b488
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4977682
Reviewed-by: Caroline Rising <corising@chromium.org>
Commit-Queue: Juan Mojica <juanmojica@google.com>
Reviewed-by: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1218360}