[iOS] Clean Up UIButtonConfiguration in Content Suggestion
Clean up for UIButtonConfigurationfeature flag. Creating small CLs to
add transparency to relevant code owners. To ensure that there
aren't conflicting UI elements, we are asking the reviewers to look at
the files in the CL and any relevant files where the modified
UIButtons are seen in other UI components. If any relevant files are
missing that the feature guard IsUIButtonConfigurationEnabled(),
please leave a comment to add those files. Feel free to reach out if
this is confusing.

Bug: 1423432
Change-Id: I0096bd54f6aa4017b59a452afbafcd5215f8307d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4995347
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Adam Arcaro <adamta@google.com>
Commit-Queue: Joemer Ramos <joemerramos@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218440}