Swap animation for promise app into installed app
This CL attempts to coordinate an animation between a promise app that
is being removed because it was successfully installed and its
corresponding pinned app in the shelf.

When a ShelfItem is removed from the ShelfModel, duplicate its layer
before removing it. ShelfView will own the duplicate layer in order to
have access to it later when the installed app is added to the shelf.

When the shelf view receives an updated shelf item, check if it has the
same package_id as any of the apps that are waiting for removal. In the
case that there is, Introduce a fade-out animation for the duplicated
layer of the promise app while the incoming app is fading in.

Bug: b:297866814
Change-Id: Ic9717f353fc9253b855282c1f58736f7c7a830f5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4936631
Commit-Queue: Ana Salazar <anasalazar@chromium.org>
Reviewed-by: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218445}