[segmentation] Add an experimental database API
The clients who cannot use segmentation execution can use the database
to store and retrieve results for model execution.
The CL adds a new DatabaseClient class through which clients can get
data as processed features from the databases. The clients need to
register the list of signals as part of a default model with this, since
the platform needs to track the list of signals stored in database for observation and cleanup. This should ideally be moved to a nicer
location in the future. The API also allows adding custom signals to the
SQL database.

Change-Id: I6564853d82d2fa276ef3bf132d2874983f9efdd0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4990220
Reviewed-by: Ritika Gupta <ritikagup@google.com>
Commit-Queue: Siddhartha S <ssid@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218452}