Run all test types by default for web platform tests
Doing this allows developers to run tests under a directory
without specifying the test type. When we need to skip certain
type of tests on a builder, do that in the builder config.

Did not make change for chrome ios because it now runs tests
from explicitly specified directories.

Bug: 1498195
Change-Id: I8fb4b83e47532c669e9e248b2f84b08f441e3652
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4997111
Commit-Queue: Weizhong Xia <weizhong@google.com>
Reviewed-by: Jonathan Lee <jonathanjlee@google.com>
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/main@{#1218456}