Add feature for browser context type
Add a feature to control the SharedMainThreadContextProvider() decoder
type from VizProcessTransportFactory. This context is currently based on
GLES2 decoder but the flag switches it to raster decoder. The feature is
default disabled now, so this is a no-op, but allows testing the change
in release builds. The feature will be default enabled as a kill switch
in a follow up CL.

Bug: 1453633
Change-Id: I4959110ecefffd9ad0339e76bab97ff06366bea8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4998234
Commit-Queue: Kyle Charbonneau <kylechar@chromium.org>
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218457}