Remove uses of implicit conversion of ScopedTypeRef in /headless
Implicit unwrapping of a scoper to its underlying pointer is dangerous
and that capability is being removed. This converts uses of implicit
conversion to be explicit so that implicit unwrapping can be removed.
This also may perform other cleanup and modernization.

This was previously landed as https://crrev.com/c/4980610 but was
reverted. This version fixes an issue.

Bug: 1495439
Change-Id: Ic6647d42005d455c35b04baa698d3693ddd13f15
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994141
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Auto-Submit: Avi Drissman <avi@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Dmitry Gozman <dgozman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218498}