[CrOS Cellular] Fix incorrect global customApnList mutation
Fix <apn-list> modifying the shared managedProperties.customApnList
when computing APNs to display to the UI. This was incorrectly removing
the connected APN from the source of truth custom APN list, causing
other methods and UI elements to have the incorrect custom APN list,
such as the <apn-detail-dialog>. This was causing <apn-detail-dialog>
to incorrectly display the "Default APN required" warning even if there
is a default APN, if it happens to be connected. This also caused the
error toasts to incorrectly be shown for the same reason.

Bug: b:308480620
Test: *ApnList*
Change-Id: I27369237490f69766db92e757a7f3c032fb6fe39
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4998117
Auto-Submit: Gordon Seto <gordonseto@google.com>
Commit-Queue: Gordon Seto <gordonseto@google.com>
Reviewed-by: Regan Hsu <hsuregan@chromium.org>
Commit-Queue: Regan Hsu <hsuregan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218509}