[Read Anything] Introduce a new experiment showing an icon in the omnibox.
This experiment is intended to improve discoverability of the feature.
When a user opens a distillable page (i.e. a page where reading mode
is expected to perform well), a reading mode icon will appear in the
omnibox, aside the zoom button. When a user clicks on the icon, the
side panel will open with reading mode visible.

I will be submitting a few follow-up CLs, with TODOs left in this CL
marking what updates are needed.

Bug: 1266555
Change-Id: I4642a521dc2794b830681df5c41c861208e09a3d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4943818
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1218512}