[E2E] Fullscreen check in DisplayCutoutController
This fixes a regression in the DCC that caused it to be active even
when not in fullscreen mode anytime the EdgeToEdge Feature is enabled.

Adds a test for this regression.

BUG: 1498624,1475820
Change-Id: I096db8787370aca1ba7ed9026eb0a2c70e9aa575
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994746
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Commit-Queue: Donn Denman <donnd@chromium.org>
Commit-Queue: Theresa Sullivan <twellington@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1218544}