Pipe shortcut app badge to shelf model
For app shortcuts, app items in the shelf should have a badge (the
badge will be the app icon of the app that hosts the app shortcut).
This updates AppIconLoaderDelegate::OnAppImageUpdated to pass the badge
image in addition to the app image. The badge image is currently set for
app shortcuts only.

BUG=b:306294395

Change-Id: I3ef60004c924e268c94a7e28e5967aa999282239
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4987448
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Reviewed-by: Giovanni Ortuno Urquidi <ortuno@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218589}