[signin] Add pixel tests for signed in profile menu
We already have pixel tests for different states when the user is not
signed in. This adds pixel tests for the states that the profile menu
can have when the user is signed in.
This test also updates existing screenshots for the following tests, as
a profile picture is added:
- EnterpriseWelcomeUIDialogPixelTest
- EnterpriseWelcomeUIWindowPixelTest

Bug: b:303415175

Change-Id: I9668309ae42763f90c77d8503337c7da00690047
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4973965
Commit-Queue: Amelie Schneider <amelies@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Jack Yammine <jyammine@google.com>
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218707}