[Exo] Update window radii sent to the clients
In the CL, crrev.com/c/4813985, we add an event that send window radii
to the clients. The protocol send non-zero radius for all four corners
regardless of rounded windows feature.

This CL, the a non-zero radius for bottom two corners if the rounded
window feature is enabled.

Bug: b:294287245
Change-Id: I4a8b7a0f4d6fbf1bf041703a345845c0ee55626b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4989857
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Zoraiz Naeem <zoraiznaeem@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218710}