Fix LevensteinDistance()
LevensteinDistance("ananas", "banana") returns 4 instead of 2. This is
because the algorithm doesn't handle the case where i >= 0 and j == -1
correctly (due to 0-based indexing j == -1 corresponds to the empty
prefix of b). In this case, the distance should be i+1, the length of
the prefix of a.

Bug: 1453650
Change-Id: Ib62eceee3d4518c2208f08c3a99f744d3b9e1805
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4987428
Reviewed-by: Christoph Schwering <schwering@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1218718}