Fix clearing generated password suggestions
For forms with 2 password fields (eg. password reset) clearing previewed
form starts by checking if it can clear autofill selection. That only
clears the username and password fields. In the case of password reset
forms that function will return true since there was a suggested
password and it was cleared. Because of that, the call to the function
DidClearGenerationSuggestion never happens, meaning that password
confirmation field is never cleared.

Bug: 1484514
Change-Id: I03ec6b3bc3fe8f3a66e40b1b0db94c86dc052455
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994384
Reviewed-by: Dominic Battre <battre@chromium.org>
Commit-Queue: Karol Sygiet <sygiet@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Maria Kazinova <kazinova@google.com>
Cr-Commit-Position: refs/heads/main@{#1218745}