[iOS] Remove PermissionsInfobarModalOverlayMediator as WebStateObserver
This makes PermissionsInfobarModalOverlayMediator remove itself as
a WebStateObserver when the WebState it is observing is about to
be destroyed.

Failing to remove itself as an observer in this situation eventually
leads to a DumpWithoutCrashing when the WebState is destroyed, since
the WebState's ObserverList should be empty by this point.

Change-Id: Id373341e90d3672c1f3302b6933ebe27946add0a
Bug: 1491782
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4995607
Commit-Queue: Ali Juma <ajuma@chromium.org>
Reviewed-by: Ginny Huang <ginnyhuang@chromium.org>
Reviewed-by: Chris Lu <thegreenfrog@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1218755}