kiosk: Refactor recovery flow
The recovery flow is currently invoked through LaunchAppOrDie. This
leads to the AppLaunchManager being leaked, which causes test failures
when switching to the WebKioskAppServiceLauncher. This CL refactors this
flow in a way that the ownership can be taken by an outside class and
therefore can be cleanly destroyed during shutdown.

Change-Id: I8f3be87266e699dca6d78ee2289e6717e7bd9856
Bug: b:242023891
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4994484
Commit-Queue: Ben Franz <bfranz@chromium.org>
Reviewed-by: Edman Anjos <edman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218776}