[autopip] Ensure that the autopip setting widget closes when pip hides
This CL makes the video picture-in-picture window remove and delete the
AutoPipSettingOverlayView when it is hidden, and modifies the
AutoPipSettingOverlayView destructor to ensure that its BubbleWidget
gets closed. This prevents an issue where mutliple
AutoPipSettingOverlayViews are shown at once on a video
picture-in-picture window when it hides and re-shows multiple times.

Bug: 1498772
Change-Id: I54c6dea87e493b78d10512d3bfd038b35d0d4264
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4999108
Reviewed-by: Fr <beaufort.francois@gmail.com>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218795}