Make IBAN functions not pure virtual in AutofillClient
This CL makes IBAN related functions in AutofillClient non-virtual. This
is a cleaner implementation and we should go in this direction for
payments features as many (if not all) of them aren't implemented on all
platforms, resulting in us having a lot of overridden functions that
aren't implemented or BUILDFLAGs to protect against possible not-
implemented compile errors.

This CL is a pre-requisite for go/autofillclient-payments-refactor-design.

Bug: 1494289
Change-Id: I43c91a4c2e9eccdc259c25afc3f2c93ac6d645c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4974863
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Mohamed Amir Yosef <mamir@chromium.org>
Commit-Queue: Vinny Persky <vinnypersky@google.com>
Cr-Commit-Position: refs/heads/main@{#1218803}