[LongClick] Add long click to all nodes if feature flag is enabled
This CL implements the main feature work of the experiment to include
long click actions on the web. We also include unit tests.

AX-Relnotes: Note for developers of AT: When the 'AccessibilityIncludeLongClickAction' feature flag is enabled, all nodes on the web will include the ACTION_LONG_CLICK in the action list of their AccessibilityNodeInfo object.
Bug: b/274532031
Change-Id: Ic44ad78c4009f6cc6afed34bf6b01c84048187da
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4997391
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Reviewed-by: Abigail Klein <abigailbklein@google.com>
Cr-Commit-Position: refs/heads/main@{#1218828}