Re-enable ChromeNetworkServiceBrowserTest.EncryptedCookies test on macOS
In crrev.com/c/4995700 a race condition was fixed whereby the
same network context directory was being shared between the
profile and the test network context.

This should hopefully fix the flakes on macOS so this CL
re-enables the test on that platform.

Bug: 868667

Include-Ci-Only-Tests: false
Change-Id: Ib142e57dab6484d1a5127132161a26e95d885ab2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4998085
Commit-Queue: Will Harris <wfh@chromium.org>
Reviewed-by: Adam Rice <ricea@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218862}