WebView: Add SafeModeAction Name Logging
WebView has multiple SafeMode actions that can be deployed to devices to
recover to a stable state. In the event that one SafeMode action is
deployed in an attempt to recover to a stable state and later is deemed
to be an incorrect choice for recovery, another SafeMode action may be
deployed. In this case, it is useful to monitor which SafeMode action is
deployed to the population and helps determine if the second action is
also utilized.

Test: out/Default/bin/run_webview_instrumentation_test_apk -f SafeModeTest*
Bug: 1473738
Change-Id: I013a4ad98236ae07f3768ee9dd3c758b348ef571
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4789456
Reviewed-by: Nate Fischer <ntfschr@chromium.org>
Commit-Queue: Adam Walls <avvall@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218889}