LinkPreview: Remove PreviewTestHeler::ActivateForTesting
This patch removes  PreviewTestHelper::ActivateForTesting as it was
tentative solution until the PromoteToNewTab is implemented.

Also this change reminds me that the WebContentsDelegate is removed
after starting the tab promotion, and no one could check the
PreviewTab::Status. So, we just assume the page is in the preview
mode while the PreviewTab specific WebContentsDelegate is alive.

BUG: b/305012045
Change-Id: I60640fd946294caf91aae4e0206461ef3cd7023a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4997413
Reviewed-by: Ken Okada <kenoss@chromium.org>
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218912}