Add WebContentsObserver::OnFrameIsCapturingVideoStreamChanged()
This new method will allow observers to get notified when a frame
is currently capturing a video stream. This will be used to move the
logic of render process backgrounding to the performance manager.

To do this, the VideoCaptureHost is changed to be associated with
the hosting RenderFrameHost, instead of the RenderProcessHost.

Bug: 1077217
Change-Id: I4bb84632fa87390fd6409c8d04c155c21994c00a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4907955
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Guido Urdaneta <guidou@chromium.org>
Reviewed-by: Brendon Tiszka <tiszka@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1218916}